+++ src/test/java/com/silverbars/LiveOrderBoardTest.java
@@ -134,11 +134,11 @@ public class LiveOrderBoardTest {
 
     // a tiny DSL to improve the readability of the test scenarios
 
-    private static OrderSummary buyOrderSummary(@kg int quantity, PricePerKg pricePerKg) {
+    private static OrderSummary buyOrderSummary(@kg double quantity, PricePerKg pricePerKg) {
         return new OrderSummary(quantity, pricePerKg, Order.Type.Buy);
     }
 
-    private static OrderSummary sellOrderSummary(@kg int quantity, PricePerKg pricePerKg) {
+    private static OrderSummary sellOrderSummary(@kg double quantity, PricePerKg pricePerKg) {
         return new OrderSummary(quantity, pricePerKg, Order.Type.Sell);
     }
 }
